assignment2/requirements.txt
click==8.1.8
glfw==2.8.0
multipledispatch==1.0.0
numpy==1.26.4
opencv-python==4.10.0.84
pyopengl==3.1.7
pyopengl-accelerate==3.1.7
pyrr==0.10.3
openmesh==1.2.1